feat: providing a simpler hook example
